backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
sqlparse==0.4.4
mysql-connector-python==8.3.0
mysqlclient==2.2.1
pytz
psycopg2-binary
